Split MarkdownDocumentUtilities: pure AST/regex helpers into Core
Batch 2f. About 150 pure lines out of 1168. Verified each of the
plan's listed members individually against invariant 8 rather than
trusting the list; all checked out as pure Markdig/regex/string code
with no FlowDocument or other System.Windows.Documents types:
LooksLikeMarkdown, ShouldPromoteLiveBlock, ShouldPromoteLiveMarkdown,
NormalizeDocumentText, NormalizeNewlines, EscapeMarkdownText,
EscapeLinkDestination, ApplyQuotePrefix, GetQuotePrefix,
GetOrderedListStart, ResolveContentSpan, GetSourceSlice,
GetCodeSpanContentRawStart, GetCodeBlockText, the MarkdownPipeline
field, and the three [GeneratedRegex] partial methods.

A partial class cannot span assemblies, so the split needed a new
type name for the impure half, per the PatternItem/PatternItemCatalog
shape from e677b54:

- Text-Grab.Core/Utilities/MarkdownDocumentUtilities.cs keeps the
  original type name and namespace (Text_Grab.Utilities), holding only
  the members above. The three now-private-in-Core regex-backed static
  fields (LiveBlockTriggerRegex etc.) moved with the methods that use
  them. Previously-private members that the app-side facade still
  calls became `internal` (MarkdownPipeline field included) - Core
  already has InternalsVisibleTo("Text-Grab") from e677b54, so no new
  visibility plumbing was needed.

- Text-Grab/Utilities/MarkdownFlowDocumentUtilities.cs (renamed from
  the original file) keeps everything FlowDocument-bound: CreateFlowDocument,
  SerializeToMarkdown, GetDocumentPlainText, BuildOffsetMap,
  MapRawOffsetToPosition and their private helpers, ApplyTheme, the
  AppendBlock/WriteBlock tree, and all the attached DependencyProperty
  plumbing (its typeof() owner arguments now point at the new type,
  since that's the class actually declaring them). Every call from
  here into the moved pure helpers is qualified with
  `MarkdownDocumentUtilities.`.

Call sites fixed (all compiler-flagged after the rename, no
unrequested edits): Text-Grab/Views/EditTextWindow.xaml.cs - the three
pure calls (ShouldPromoteLiveMarkdown, ShouldPromoteLiveBlock,
LooksLikeMarkdown) keep referring to MarkdownDocumentUtilities; the
FlowDocument-touching calls (CreateFlowDocument, BuildOffsetMap,
SerializeToMarkdown, ApplyTheme, MapRawOffsetToPosition,
GetDocumentPlainText, the MarkdownOffsetMap field type) were renamed
to MarkdownFlowDocumentUtilities.

Markdig stays referenced by both halves, as the plan calls out: the
app side pattern-matches on Markdig AST types directly in
AppendBlock/AppendInline, so its own PackageReference had to stay -
verified this is a real compile need, not just relying on
transitivity, before leaving it in place. Added the same Markdig
1.3.2 PackageReference to Text-Grab.Core.csproj so the pure half's
Markdig.Syntax/Markdig.Syntax.Inlines/Markdig.Extensions.AutoIdentifiers
usages resolve.

Extracted the 6 pure test methods (all [Theory], no WpfFact, no
FlowDocument) into Tests.Core/MarkdownParsingTests.cs, namespace
Text_Grab.Tests.Core matching the ThirdPartyNoticeUtilitiesTests
convention. Renamed the remaining 33 FlowDocument-dependent call sites
in Tests/MarkdownDocumentUtilitiesTests.cs from
MarkdownDocumentUtilities to MarkdownFlowDocumentUtilities;
mechanical, no other changes.

Test counts, migration shift accounted for:
- Tests.Core: 487 -> 514 (+27, the 6 Theory methods' InlineData cases:
  6+3+5+4+6+3)
- Tests.Core.Windows: 3 -> 3 (unchanged)
- Tests: 1020 -> 993 passed / 7 skipped (-27 passed, migrated out)
27 + 993 = 1020, so no regression hid inside the shift.
